+++ homeassistant/components/todoist/calendar.py
@@ -3,7 +3,6 @@
 from datetime import date, datetime, timedelta
 import logging
 from typing import Any, override
-import uuid
 
 from todoist_api_python.api_async import TodoistAPIAsync
 from todoist_api_python.models import Label, Project, Task
@@ -15,10 +14,8 @@ from homeassistant.components.calendar import (
     CalendarEvent,
 )
 from homeassistant.const import CONF_ID, CONF_NAME, CONF_TOKEN, EVENT_HOMEASSISTANT_STOP
-from homeassistant.core import Event, HomeAssistant, ServiceCall, callback
-from homeassistant.exceptions import ServiceValidationError
+from homeassistant.core import Event, HomeAssistant, callback
 from homeassistant.helpers import config_validation as cv
-from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.entity_platform import (
     AddConfigEntryEntitiesCallback,
     AddEntitiesCallback,
@@ -30,30 +27,17 @@ from homeassistant.util import dt as dt_util
 from .const import (
     ALL_DAY,
     ALL_TASKS,
-    ASSIGNEE,
     COMPLETED,
     CONF_EXTRA_PROJECTS,
     CONF_PROJECT_DUE_DATE,
     CONF_PROJECT_LABEL_WHITELIST,
     CONF_PROJECT_WHITELIST,
-    CONTENT,
     DESCRIPTION,
-    DOMAIN,
-    DUE_DATE,
-    DUE_DATE_LANG,
-    DUE_DATE_STRING,
-    DUE_DATE_VALID_LANGS,
     DUE_TODAY,
     END,
     LABELS,
     OVERDUE,
     PRIORITY,
-    PROJECT_NAME,
-    REMINDER_DATE,
-    REMINDER_DATE_LANG,
-    REMINDER_DATE_STRING,
-    SECTION_NAME,
-    SERVICE_NEW_TASK,
     START,
     SUMMARY,
 )
@@ -63,25 +47,6 @@ from .util import parse_due_date
 
 _LOGGER = logging.getLogger(__name__)
 
-NEW_TASK_SERVICE_SCHEMA = vol.Schema(
-    {
-        vol.Required(CONTENT): cv.string,
-        vol.Optional(DESCRIPTION): cv.string,
-        vol.Optional(PROJECT_NAME, default="inbox"): vol.All(cv.string, vol.Lower),
-        vol.Optional(SECTION_NAME): vol.All(cv.string, vol.Lower),
-        vol.Optional(LABELS): cv.ensure_list_csv,
-        vol.Optional(ASSIGNEE): cv.string,
-        vol.Optional(PRIORITY): vol.All(vol.Coerce(int), vol.Range(min=1, max=4)),
-        vol.Exclusive(DUE_DATE_STRING, "due_date"): cv.string,
-        vol.Optional(DUE_DATE_LANG): vol.All(cv.string, vol.In(DUE_DATE_VALID_LANGS)),
-        vol.Exclusive(DUE_DATE, "due_date"): cv.string,
-        vol.Exclusive(REMINDER_DATE_STRING, "reminder_date"): cv.string,
-        vol.Optional(REMINDER_DATE_LANG): vol.All(
-            cv.string, vol.In(DUE_DATE_VALID_LANGS)
-        ),
-        vol.Exclusive(REMINDER_DATE, "reminder_date"): cv.string,
-    }
-)
 
 PLATFORM_SCHEMA = CALENDAR_PLATFORM_SCHEMA.extend(
     {
@@ -127,7 +92,6 @@ async def async_setup_entry(
         entities.append(TodoistProjectEntity(coordinator, project_data, labels))
 
     async_add_entities(entities)
-    async_register_services(hass, coordinator)
 
 
 async def async_setup_platform(
@@ -204,157 +168,6 @@ async def async_setup_platform(
         )
 
     async_add_entities(project_devices, update_before_add=True)
-
-    async_register_services(hass, coordinator)
-
-
-def async_register_services(  # noqa: C901
-    hass: HomeAssistant, coordinator: TodoistCoordinator
-) -> None:
-    """Register services."""
-
-    if hass.services.has_service(DOMAIN, SERVICE_NEW_TASK):
-        return
-
-    session = async_get_clientsession(hass)
-
-    async def handle_new_task(call: ServiceCall) -> None:
-        """Call when a user creates a new Todoist Task from Home Assistant."""
-        project_name = call.data[PROJECT_NAME]
-        projects = await coordinator.async_get_projects()
-        project_id: str | None = None
-        for project in projects:
-            if project_name == project.name.lower():
-                project_id = project.id
-                break
-        if project_id is None:
-            raise ServiceValidationError(
-                translation_domain=DOMAIN,
-                translation_key="project_invalid",
-                translation_placeholders={
-                    "project": project_name,
-                },
-            )
-
-        # Optional section within project
-        section_id: str | None = None
-        if SECTION_NAME in call.data:
-            section_name = call.data[SECTION_NAME]
-            sections = await coordinator.async_get_sections(project_id)
-            for section in sections:
-                if section_name == section.name.lower():
-                    section_id = section.id
-                    break
-            if section_id is None:
-                raise ServiceValidationError(
-                    translation_domain=DOMAIN,
-                    translation_key="section_invalid",
-                    translation_placeholders={
-                        "section": section_name,
-                        "project": project_name,
-                    },
-                )
-
-        # Create the task
-        content = call.data[CONTENT]
-        data: dict[str, Any] = {"project_id": project_id}
-
-        if description := call.data.get(DESCRIPTION):
-            data["description"] = description
-
-        if section_id is not None:
-            data["section_id"] = section_id
-
-        if task_labels := call.data.get(LABELS):
-            data["labels"] = task_labels
-
-        if ASSIGNEE in call.data:
-            collaborators_result = await coordinator.api.get_collaborators(project_id)
-            all_collaborators = await flatten_async_pages(collaborators_result)
-            collaborator_id_lookup = {
-                collab.name.lower(): collab.id for collab in all_collaborators
-            }
-            task_assignee = call.data[ASSIGNEE].lower()
-            if task_assignee in collaborator_id_lookup:
-                data["assignee_id"] = collaborator_id_lookup[task_assignee]
-            else:
-                raise ValueError(
-                    f"User is not part of the shared project. user: {task_assignee}"
-                )
-
-        if PRIORITY in call.data:
-            data["priority"] = call.data[PRIORITY]
-
-        if DUE_DATE_STRING in call.data:
-            data["due_string"] = call.data[DUE_DATE_STRING]
-
-        if DUE_DATE_LANG in call.data:
-            data["due_lang"] = call.data[DUE_DATE_LANG]
-
-        if DUE_DATE in call.data:
-            due_date = dt_util.parse_datetime(call.data[DUE_DATE])
-            if due_date is None:
-                due = dt_util.parse_date(call.data[DUE_DATE])
-                if due is None:
-                    raise ValueError(f"Invalid due_date: {call.data[DUE_DATE]}")
-                due_date = datetime(due.year, due.month, due.day)
-            # Pass the datetime object directly - the library handles formatting
-            data["due_datetime"] = dt_util.as_utc(due_date)
-
-        api_task = await coordinator.api.add_task(content, **data)
-
-        # The REST API doesn't support reminders, so we use the Sync API directly
-        # to maintain functional parity with the component.
-        # https://developer.todoist.com/api/v1/#tag/Sync/Reminders/Add-a-reminder
-        _reminder_due: dict = {}
-        if REMINDER_DATE_STRING in call.data:
-            _reminder_due["string"] = call.data[REMINDER_DATE_STRING]
-
-        if REMINDER_DATE_LANG in call.data:
-            _reminder_due["lang"] = call.data[REMINDER_DATE_LANG]
-
-        if REMINDER_DATE in call.data:
-            reminder_date = dt_util.parse_datetime(call.data[REMINDER_DATE])
-            if reminder_date is None:
-                reminder = dt_util.parse_date(call.data[REMINDER_DATE])
-                if reminder is None:
-                    raise ValueError(
-                        f"Invalid reminder_date: {call.data[REMINDER_DATE]}"
-                    )
-                reminder_date = datetime(reminder.year, reminder.month, reminder.day)
-            # Format it in the manner Todoist expects (UTC with Z suffix)
-            reminder_date = dt_util.as_utc(reminder_date)
-            date_format = "%Y-%m-%dT%H:%M:%S.000000Z"
-            _reminder_due["date"] = datetime.strftime(reminder_date, date_format)
-
-        if _reminder_due:
-            sync_url = "https://api.todoist.com/api/v1/sync"
-            reminder_data = {
-                "commands": [
-                    {
-                        "type": "reminder_add",
-                        "temp_id": str(uuid.uuid1()),
-                        "uuid": str(uuid.uuid1()),
-                        "args": {
-                            "item_id": api_task.id,
-                            "type": "absolute",
-                            "due": _reminder_due,
-                        },
-                    }
-                ]
-            }
-            headers = {
-                "Authorization": f"Bearer {coordinator.token}",
-                "Content-Type": "application/json",
-            }
-            await session.post(sync_url, headers=headers, json=reminder_data)
-
-        _LOGGER.debug("Created Todoist task: %s", call.data[CONTENT])
-
-    # pylint: disable-next=home-assistant-service-registered-in-setup-entry
-    hass.services.async_register(
-        DOMAIN, SERVICE_NEW_TASK, handle_new_task, schema=NEW_TASK_SERVICE_SCHEMA
-    )
 
 
 class TodoistProjectEntity(CoordinatorEntity[TodoistCoordinator], CalendarEntity):
